feat: consolidate RSS tests into builders.test.js, fix buildRss async and CORS proxy
