+++ src/config.json
@@ -1,1 +1,31 @@
-{}
+{
+  "proxy": "https://api.allorigins.win/raw?url=",
+  "layout": {
+    "zones": ["header", "main", "sidebar", "footer"]
+  },
+  "theme": {
+    "background": "#111111",
+    "color": "#ffffff",
+    "fontFamily": "sans-serif"
+  },
+  "components": [
+    {
+      "zone": "sidebar",
+      "type": "image",
+      "src": "./assets/logo.png",
+      "alt": "Your image description here"
+    },
+    {
+      "zone": "main",
+      "type": "clock",
+      "mode": "digital"
+    },
+    {
+      "zone": "main",
+      "type": "rss",
+      "url": "https://feeds.bbci.co.uk/news/rss.xml",
+      "maxItems": 5,
+      "refresh": 60000
+    }
+  ]
+}
